preserve enclosing [] to ensure ipv6 lit hostname remains valid
In the case whre hostname arrives as a bracketed ipv6 literal,
we don't split into separate parts based on '.' as that's not valid
within an ipv6 address. In these cases, this change lets us preserve
the address in a way that clients can use directly, without having to
check for ipv6/ipv4/hostname formatting.