Match results of file and log operations
This patch resolves dialyzer warnings for unmatched returns. The
behavior of the logging modules is changed: if the low level file
operations fail, the calling process will crash with a bad match error
whereas previously it would carry on. For all of the cases, crashing
seems like the right thing since failure to create a timer or write a
log file should not be ignored.